third_party/WebKit/LayoutTests/typedcssom/inlinestyle/properties/transform-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = 2 duplicate test names: "Setting transform to CSSTransformValue with value ", "Getting transform with a CSSTransformValue whose value is "
PASS Setting transform to none 
PASS Setting transform to initial 
PASS Setting transform to inherit 
PASS Setting transform to unset 
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting transform to CSSTransformValue with value  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
PASS Setting transform to invalid value CSSUnitValue "4px" throws 
PASS Setting transform to invalid value null throws 
PASS Setting transform to invalid value undefined throws 
PASS Setting transform to invalid value true throws 
PASS Setting transform to invalid value false throws 
PASS Setting transform to invalid value 1 throws 
PASS Setting transform to invalid value hello throws 
PASS Setting transform to invalid value [object Object] throws 
PASS Setting transform to invalid value CSSKeywordValue "notAKeyword" throws 
PASS Getting transform when it is set to none 
PASS Getting transform when it is set to initial 
PASS Getting transform when it is set to inherit 
PASS Getting transform when it is set to unset 
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform with a CSSTransformValue whose value is  Cannot read property 'constructor' of null
FAIL Getting transform when it is set to "skew(45deg)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skew(1rad)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skew(6.2grad)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skew(0.31turn)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skewX(45deg)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skewX(1rad)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skewX(6.2grad)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skewX(0.31turn)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skewY(45deg)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skewY(1rad)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skewY(6.2grad)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
FAIL Getting transform when it is set to "skewY(0.31turn)" via a string assert_equals: typeof result expected "CSSTransformValue" but got "CSSStyleValue"
PASS Getting transform when it is set to "rotateX(45deg)" via a string 
PASS Getting transform when it is set to "rotateX(1rad)" via a string 
PASS Getting transform when it is set to "rotateX(6.2grad)" via a string 
PASS Getting transform when it is set to "rotateX(0.31turn)" via a string 
PASS Getting transform when it is set to "rotateY(45deg)" via a string 
PASS Getting transform when it is set to "rotateY(1rad)" via a string 
PASS Getting transform when it is set to "rotateY(6.2grad)" via a string 
PASS Getting transform when it is set to "rotateY(0.31turn)" via a string 
PASS Getting transform when it is set to "rotateZ(45deg)" via a string 
PASS Getting transform when it is set to "rotateZ(1rad)" via a string 
PASS Getting transform when it is set to "rotateZ(6.2grad)" via a string 
PASS Getting transform when it is set to "rotateZ(0.31turn)" via a string 
FAIL getAll for single-valued transform assert_equals: Expected getAll to retrieve an array containing a single CSSStyleValue expected 1 but got 0
FAIL Delete transform removes the value from the styleMap assert_not_equals: got disallowed value null
FAIL transform shows up in getProperties Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
PASS Setting transform to a sequence throws 
PASS Appending to transform throws 
PASS Unhandled case doesn't crash. 
PASS Getting transform when it has a rotate with a calc angle does not crash 
PASS Getting transform when it has a rotate3d with a calc angle does not crash 
PASS Getting transform when it has a skew with a calc angle does not crash 
PASS Getting transform when it has a 2-argument skew with a calc angle does not crash 
PASS Getting transform when it has a skewX with a calc angle does not crash 
PASS Getting transform when it has a skewY with a calc angle does not crash 
Harness: the test ran to completion.
